+++ NEWS.md
@@ -3,6 +3,7 @@ shinyWidgets 0.5.4
 
 * New functions: `currencyInput()`, `formatNumericInput()` and `autonumericInput()` to enter numeric value with specified format like a currency, by [@srmatth](https://github.com/srmatth).
 * Ability to update icons in `updateNumericInputIcon()` and `updateTextInputIcon()`.
+* SweetAlert family: upgraded to 9.17.1 and added a polyfill to work in Internet Explorer (see `useSweetAlert()`).
 
 
 ### Bug fixes